Mostly finish

+++ main.go
@@ -1,12 +1,13 @@
 package main
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strconv"
+
+	"github.com/schollz/progressbar/v3"
 )
 
 const cachefolder = "buildprog"
@@ -27,17 +28,26 @@ func main() {
 	cachedir, err := os.UserCacheDir()
 	handle(err)
 
-	if os.Args[1] == "-h" || os.Args[1] == "--help" {
-		fmt.Println("buildprog [-cleancache] Go compiler arguments")
+	start := 1
+	// Flags
+	if len(os.Args) > 1 {
+		if os.Args[1] == "-h" || os.Args[1] == "--help" {
+			fmt.Println("buildprog [-cleancache] Go compiler arguments")
+			return
+		}
+
+		if os.Args[1] == "-cleancache" {
+			fmt.Println("Removing cache...")
+			err = os.RemoveAll(filepath.Join(cachedir, cachefolder))
+			handle(err)
+			start++
+		}
 	}
 
-	start := 1
-	if os.Args[1] == "-cleancache" {
-		err = os.RemoveAll(filepath.Join(cachedir, cachefolder))
-		handle(err)
-		start++
+	flags := make([]string, 0)
+	if len(os.Args) > start {
+		flags = os.Args[start:]
 	}
-	flags := os.Args[start:]
 
 	path := filepath.Join(cachedir, cachefolder, currdir)
 	err = os.MkdirAll(path, os.ModePerm)
@@ -46,19 +56,37 @@ func main() {
 	outFile := filepath.Join(path, buildFile)
 	cmd := exec.Command("go", append(buildArgs, flags...)...)
 
-	errout := bytes.NewBufferString("")
-	cmd.Stderr = errout
-
 	if _, err := os.Stat(outFile); os.IsNotExist(err) {
 		fmt.Println("No analysis graph available, compiling without progress this time...")
-		out, err := os.Create(outFile)
+
+		buff := &lenBuff{}
+		cmd.Stderr = buff
+		err = cmd.Run()
 		handle(err)
-		cmd.Stdout = out
+
+		err = os.WriteFile(outFile, []byte(strconv.Itoa(buff.len)), os.ModePerm)
+		handle(err)
 	} else {
-		fmt.Println("unimplemented!")
+		fmt.Println("Loading cache...")
+		cache, err := os.ReadFile(outFile)
+		handle(err)
+		prevlen, err := strconv.Atoi(string(cache))
+		handle(err)
+
+		bar := progressbar.New(prevlen)
+		buff := &progBuff{
+			origLen: prevlen,
+			bar:     bar,
+			len:     0,
+		}
+
+		fmt.Println("Building...")
+
+		cmd.Stderr = buff
+		err = cmd.Run()
+		handle(err)
+
+		buff.Close(outFile)
 	}
-	err = cmd.Run()
-	if err != nil {
-		handle(errors.New(errout.String()))
-	}
+	fmt.Println("Successfully built!")
 }
